acceptance: cleanup pass
