+++ acceptance/common/base.py
@@ -16,15 +16,14 @@ import logging
 import os
 import re
 import traceback
-from typing import List
 
 from plumbum import cli
 from plumbum import local
 from plumbum import cmd
 from plumbum import LocalPath
 
-from acceptance.common.docker import Compose
-from acceptance.common import log
+from acceptance.common import docker, log
+from tools.topology.scion_addr import ISD_AS
 
 logger = logging.getLogger(__name__)
 
@@ -50,7 +49,8 @@ def ContainerLoader(arg):
 
 class TestBase:
     # XXX(matzf) should be executable?
-    @cli.switch("executables", NameExecutable, list=True, help="Paths for executables, format name:path")
+    @cli.switch("executables", NameExecutable, list=True,
+                help="Paths for executables, format name:path")
     def _set_executables(self, executables):
         self.executables = {name: executable for (name, executable) in executables}
 
@@ -79,6 +79,7 @@ class TestBase:
     def setup_prepare(self):
         """Unpacks loads local docker images and generates the topology.
         """
+        docker.assert_no_networks()
         self._setup_artifacts()
         self._setup_container_loaders()
         # Define where coredumps will be stored.
@@ -154,7 +155,7 @@ class TestTopogen(TestBase):
     def setup_start(self):
         """Starts the docker containers in the topology.
         """
-        self.dc = Compose(compose_file=self.artifacts / "gen/scion-dc.yml")
+        self.dc = docker.Compose(compose_file=self.artifacts / "gen/scion-dc.yml")
         print(self.dc("up", "-d"))
         ps = self.dc("ps")
         print(ps)
@@ -171,87 +172,13 @@ class TestTopogen(TestBase):
         if re.search(r"Exit\s+[1-9]\d*", ps):
             raise Exception("Failed services.\n" + ps)
 
-    # XXX(matzf) move all these to Compose
-    def start_container(self, container):
-        """Starts the container with the specified name.
-
-        Args:
-            container: the name of the container.
-        """
-        print(self.dc("start", container))
-
-    def restart_container(self, container):
-        """Restarts the container with the specified name.
-
-        Args:
-            container: the name of the container.
-        """
-        print(self.dc("restart", container))
-
-    def stop_container(self, container):
-        """Stops the container with specified name.
-
-        Args:
-            container: the name of the container.
-        """
-        print(self.dc("stop", container))
-
-    def list_containers(self, container_pattern: str) -> List[str]:
-        """Lists all containers that match the given pattern.
-
-        Args:
-            container_pattern: A regex string to match the container. The regex
-              format is standard Python regex format.
-
-        Returns:
-            A list of strings with the container names that match the
-            container_pattern regex.
-        """
-        containers = self.dc("config", "--services")
-        matching_containers = []
-        for container in containers.splitlines():
-            if re.match(container_pattern, container):
-                matching_containers.append(container)
-        return matching_containers
-
-    def send_signal(self, container, signal):
-        """Sends signal to the container with the specified name.
-
-        Args:
-            container: the name of the container.
-            signal: the signal to send
-        """
-        print(self.dc("kill", "-s", signal, container))
-
-    def execute(self, container, *args, **kwargs):
-        """Executes an arbitrary command in the specified container.
-
-        There's one minute timeout on the command so that tests don't get stuck.
-
-        Args:
-            container: the name of the container to execute the command in.
+    def execute_tester(self, isd_as: ISD_AS, cmd: str, *args: str) -> str:
+        """Executes a command in the designated "tester" container for the specified ISD-AS.
 
         Returns:
             The output of the command.
         """
-        user = kwargs.get("user", "{}:{}".format(os.getuid(), os.getgid()))
-        return self.dc("exec", "-T", "--user", user, container,
-                       "timeout", "1m", *args)
-
-    def execute_as_user(self, container, user, *args, **kwargs):
-        """Executes an arbitrary command in the specified container.
-
-        There's one minute timeout on the command so that tests don't get stuck.
-
-        Args:
-            container: the name of the container to execute the command in.
-            user: the user to use to execute the command
-
-        Returns:
-            The output of the command.
-        """
-        return self.dc("exec", "-T", "--user", user, container,
-                       "timeout", "1m", *args)
+        return self.dc.execute("tester_%s" % isd_as.file_fmt(), cmd, *args)
 
 
 def main(test_class):
