+++ acceptance/common/log.py
@@ -13,24 +13,6 @@
 # limitations under the License.
 
 import logging
-from functools import wraps
-
-
-class LogExec(object):
-    def __init__(self, logger: logging.Logger, sub_command: str):
-        self.sub_command = sub_command
-        self.logger = logger
-
-    def __call__(self, f):
-        @wraps(f)
-        def wrapper(*args, **kwargs):
-            self.logger.info("Start %s" % self.sub_command)
-            ret = f(*args, **kwargs)
-            if ret:
-                self.logger.warning("Failed %s" % self.sub_command)
-                return ret
-            self.logger.info("Finished %s" % self.sub_command)
-        return wrapper
 
 
 def init_log():
