+++ acceptance/trc_update/test.py
@@ -72,9 +72,9 @@ class Test(base.TestTopogen):
         end2end["-d", "-outDir", artifacts].run_fg()
 
         logger.info('==> Shutting down control servers and purging caches')
-        cs_services = self.list_containers(".*_cs.*")
+        cs_services = self.dc.list_containers(".*_cs.*")
         for cs in cs_services:
-            self.stop_container(cs)
+            self.dc.stop_container(cs)
 
         for cs_config in cs_configs:
             files = artifacts // ('gen-cache/%s*' % cs_config.stem)
@@ -83,7 +83,7 @@ class Test(base.TestTopogen):
             logger.info('Deleted files: %s' % [file.name for file in files])
 
         for cs in cs_services:
-            self.start_container(cs)
+            self.dc.start_container(cs)
         time.sleep(5)
 
         logger.info('==> Check connectivity')
